+++ tests/extensions/test_check_docs.py
@@ -33,7 +33,6 @@ from astroid import nodes
 
 from pylint.extensions.docparams import DocstringParameterChecker
 from pylint.testutils import CheckerTestCase, MessageTest
-from pylint.testutils.decorator import set_config_directly
 
 
 class TestParamDocChecker(CheckerTestCase):
@@ -129,103 +128,3 @@ class TestParamDocChecker(CheckerTestCase):
         )
         with self.assertNoMessages():
             self.checker.visit_functiondef(node)
-
-    @set_config_directly(no_docstring_rgx=re.compile(r"^_(?!_).*$"))
-    def test_skip_no_docstring_rgx(self) -> None:
-        """Example of a function that matches the default 'no-docstring-rgx' config option
-
-        No error message is emitted.
-        """
-        node = astroid.extract_node(
-            """
-        def _private_function_foo(x, y):
-            '''docstring ...
-
-            missing parameter documentation'''
-            pass
-        """
-        )
-        with self.assertNoMessages():
-            self.checker.visit_functiondef(node)
-
-    def test_all_docstring_rgx(self) -> None:
-        """Function that matches "check all functions" 'no-docstring-rgx' config option
-        No error message is emitted.
-        """
-        node = astroid.extract_node(
-            """
-        #pylint disable=missing-module-docstring, too-few-public-methods,
-        class MyClass:
-            def __init__(self, my_param: int) -> None:
-                '''
-                My init docstring
-                :param my_param: My first param
-                '''
-        """
-        )
-        with self.assertNoMessages():
-            self.checker.visit_functiondef(node.body[0])
-
-    def test_fail_empty_docstring_rgx(self) -> None:
-        """Function that matches "check all functions" 'no-docstring-rgx' config option
-        An error message is emitted.
-        """
-        node = astroid.extract_node(
-            """
-        #pylint disable=missing-module-docstring, too-few-public-methods,
-        class MyClass:
-            def __init__(self, my_param: int) -> None:
-                '''
-                My init docstring
-                '''
-        """
-        )
-        with self.assertAddsMessages(
-            MessageTest(
-                msg_id="missing-param-doc", node=node.body[0], args=("my_param",)
-            ),
-        ):
-            self.checker.visit_functiondef(node.body[0])
-
-    @set_config_directly(no_docstring_rgx=re.compile("^(?!__init__$)_"))
-    def test_fail_docparams_check_init(self) -> None:
-        """Check that __init__ is checked correctly, but other private methods aren't"""
-        node = astroid.extract_node(
-            """
-        #pylint disable=missing-module-docstring, too-few-public-methods,
-        class MyClass:
-            def __init__(self, my_param: int) -> None:
-                '''
-                My init docstring
-                '''
-
-            def _private_method(self, my_param: int) -> None:
-                '''
-                My private method
-                '''
-        """
-        )
-        with self.assertAddsMessages(
-            MessageTest(
-                msg_id="missing-param-doc", node=node.body[0], args=("my_param",)
-            )
-        ):
-            self.checker.visit_functiondef(node.body[0])
-
-    @set_config_directly(no_docstring_rgx=re.compile(""))
-    def test_no_docstring_rgx(self) -> None:
-        """Function that matches "check no functions" 'no-docstring-rgx' config option
-        No error message is emitted.
-        """
-        node = astroid.extract_node(
-            """
-        #pylint disable=missing-module-docstring, too-few-public-methods,
-        class MyClass:
-            def __init__(self, my_param: int) -> None:
-                '''
-                My init docstring
-                '''
-        """
-        )
-        with self.assertNoMessages():
-            self.checker.visit_functiondef(node.body[0])
