+++ tests/functional/ext/docparams/parameter/missing_param_doc_required_no_doc_rgx_check_init.txt
@@ -1,0 +1,1 @@
+missing-param-doc:10:4:13:11:MyClass.__init__:"""my_param"" missing in parameter documentation":UNDEFINED
