+++ tests/functional/ext/docparams/parameter/missing_param_doc_required_no_doc_rgx_default.rc
@@ -1,0 +1,6 @@
+[MASTER]
+load-plugins = pylint.extensions.docparams
+
+[BASIC]
+accept-no-param-doc=no
+docstring-min-length=-1
